requirements.txt
streamlit==1.19.0
pandas==1.5.3
pandasai==0.3.0
plotly==5.10.0
python-dotenv==1.0.0
